+++ Sources/KeyboardKit/Actions/KeyboardActionHandler.swift
@@ -45,11 +45,6 @@ public protocol KeyboardActionHandler: AnyObject, KeyboardFeedbackService {
         _ suggestion: Autocomplete.Suggestion
     )
 
-    /// Handle a certain keyboard action from a callout.
-    func handleCalloutAction(
-        _ action: KeyboardAction
-    )
-
     /// Handle a certain keyboard action drag gesture.
     func handleDrag(
         on action: KeyboardAction,